fixes issue with default files

+++ music/forms.py
@@ -38,8 +38,9 @@ class ModelEditWithRelatedFields(SmartForm, forms.Form):
                     if isinstance(new_value, InMemoryUploadedFile):
                         old_file = getattr(self.instance, key)
 
-                        if isinstance(old_file, ImageFieldFile) or isinstance(old_file, FileField):
-                            old_file.delete()
+                        if '/defaults/' not in str(old_file):
+                            if isinstance(old_file, ImageFieldFile) or isinstance(old_file, FileField):
+                                old_file.delete()
                     setattr(self.instance, key, new_value)
             if commit:
                 self.instance.save()
